refactor(component): make `DrumPad.tsx`
